+++ core/docs/changelog.md
@@ -2,6 +2,7 @@
 ### Changes
 * Upgraded to Kurt Client 0.7.0, which renames the Kurt API bindings/consts packages for easier readability in the code
 * Added CircleCI step in check_code to check for any running docker containers after Kurtosis testsuite builds and runs 
+* Depend on `kurtosis-testsuite-api-lib`, rather than `kurtosis-libs`, to get testsuite API bindings
 
 ### Fixes
 * Added check to account for error when calling Destroy method inside api_container/main.go